requirements.txt
python-dotenv
djLint
shed

flask
flask_session 
flask-socketio
flask_sqlalchemy